0.4.0开发-工作流相关接口

+++ rosettaflow-controller/src/main/java/com/moirae/rosettaflow/req/workflow/wizard/CreateWorkflowOfWizardModeReq.java
@@ -21,5 +21,5 @@ public class CreateWorkflowOfWizardModeReq {
     private Long algorithmId;
 
     @ApiModelProperty(value = "计算流程id", required = true)
-    private Integer calculationProcessId;
+    private Long calculationProcessId;
 }
