+++ rosettaflow-service/src/main/java/com/moirae/rosettaflow/service/dto/workflow/WorkflowVersionKeyDto.java
@@ -11,5 +11,5 @@ import lombok.EqualsAndHashCode;
 public class WorkflowVersionKeyDto extends WorkflowKeyDto {
 
     @ApiModelProperty(value = "工作流最新版本号")
-    private Integer workflowVersion;
+    private Long workflowVersion;
 }
